refactor: ImageProperties

+++ src/main/resources/db/migration/V2.0__create_item_Image_ddl.sql
@@ -4,6 +4,6 @@ CREATE TABLE item_image
     item_id bigint not null comment '상품 ID',
     name    VARCHAR(500) NOT NULL comment '이미지 파일 이름',
     url     VARCHAR(2500) NOT NULL comment '이미지 주소',
-    ordering tinyint(1) comment '정렬 순서',
+    ordering BIGINT comment '정렬 순서',
     FOREIGN KEY(item_id) REFERENCES item(id)
 );
